refactor resourcereader into utilities package. added jacoco gradle script. needs to be modified to work
